Hall of Heroes update
Hall of heroes has been updated and 3 new regions have been added. The
Donator's Demense is where those who donate to the game may have an icon
placed in the game. The Altar of Ancients is for those who assisted in
some way during the original release of IdleLands. The Chamber of
Contributors is for those who have helped work on the IdleLands project.
Small changes to Norkos town for the hall of heroes entrances.
Norkos Fisheries has been updated, and a collectible has been added.
Delver's rejoice!