Dungeon of Love Update
Added a new holiday dungeon, the Dungeon of Love. 2 floor dungeon with 3
new collectibles, 2 bosses, 3 boss items, 3 new regions, and a 3 piece
treasure set.

An ancient necromancer known as the Bloodkeeper has taken up residence
in the Dungeon of Love. Siphoning the power within, the Bloodkeeper uses
the harnessed energy to power his profane blood rituals. During the
month of February, all players will be available to enter this dungeon,
so prepare yourself!

Added 2 GM teleports to the current holiday maps. Note: The teleports
are to the Norkos entrances, so it still has to be during the holiday
period to enter the map this way.