Merchant's Guild Update, Ascension Map Fix, closes #117
Made the Ascension Vaults in the Merchant's Guild require the
appropriate ascension level to access the gold, loot bags, bosses, and
collectibles. The teleports are open to anyone, so a player without the
appropriate ascension level who stowsaway into the vault will have no
access to the perks, but can still escape the room.

A few items needed to be moved in the left vault to make the teleport
accessible for stowaway players.

Fixed some locked doors in the right vault, opening up 3 more treasure
bags to access.

Updated the bonuses to many of the merchants across the entire guild.