Mythos -3 Ascension Map, New Donators, and other fixes
Mythos -3 is a new Ascension Only map with new content for the level
200-250 range. You can access Mythos -3 from a new teleport located on
Mythos -1. Mythos -3 contains 6 new bosses with 5 items, and a boss
party battle to access a treasure hoard located on the map. There are
also:

10 new collectibles
A new set of treasure gear: The Old-Guard's Gear

If you make it to Mythos -3 and find some treasure or kill any bosses,
let me know! I'm looking for feedback on boss battles for high level
characters on these new high level bosses.

Next, three new players have been added to the Hall of Heroes:
xFlyingMurloc, Aeternalis, and Guts. Thank you all for your
contributions to the game!

Fixes:

Changed flavorText on the Frigri Funhouse sign, which still mentioned
needing level 20 to enter. Don't worry, everyone is welcome to enter the
maze traps of the Funhouse.

Removed boss collectibles rarities. It defaults these collectibles to
Guardian rarity.