Portal Update Part 1
This is the first part of an update to the portals of the world. Part 1
includes Norkos Overworld and the Merchants Guild, with the rest of the
maps to come soon. To standarize what you see on the map, many of the in
game portal icons will be updating: http://i.imgur.com/0VrJGQC.png

A blue portal indicates that you will move to a new map.
A red portal indicates that you will move to another location on the
same map.
A yellow portal is the special providence event located in the Fate
Pools.
A purple portal indicates a teleport that requires some level of
ascensions to enter. Hover over the teleport to find out more.
A green portal indicates a teleport location for future content for
guilds.

Bugs:
Fixed some teleporter logic in the Merchant's Guild.